crate update
admin 디자인 삭제 기능 유저 권한 수정기능완성

게임 추가 폼 완성

+++ sale/src/main/java/kr/game/sale/entity/game/Game.java
@@ -30,7 +30,7 @@ public class Game {
 
     @Id
     @Column(name ="steam_appid")
-    private int steamAppid;
+    private Long steamAppid;
 
     private String name;
